Refactoring and adding benchmarks
