+++ lib/segment/benches/metrics.rs
@@ -2,10 +2,36 @@
 mod prof;
 
 use criterion::{criterion_group, criterion_main, Criterion};
+use half::f16;
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
-use segment::data_types::vectors::VectorElementTypeByte;
+use segment::data_types::vectors::{VectorElementTypeByte, VectorElementTypeHalf};
 use segment::spaces::metric::Metric;
+#[cfg(target_arch = "x86_64")]
+use segment::spaces::metric_f16::avx::cosine::cosine_similarity_avx_half;
+#[cfg(target_arch = "x86_64")]
+use segment::spaces::metric_f16::avx::dot::dot_similarity_avx_half;
+#[cfg(target_arch = "x86_64")]
+use segment::spaces::metric_f16::avx::euclid::euclid_similarity_avx_half;
+#[cfg(target_arch = "x86_64")]
+use segment::spaces::metric_f16::avx::manhattan::manhattan_similarity_avx_half;
+#[cfg(target_arch = "aarch64")]
+use segment::spaces::metric_f16::neon::dot::dot_similarity_neon_half;
+#[cfg(target_arch = "aarch64")]
+use segment::spaces::metric_f16::neon::euclid::euclid_similarity_neon_half;
+#[cfg(target_arch = "aarch64")]
+use segment::spaces::metric_f16::neon::manhattan::manhattan_similarity_neon_half;
+use segment::spaces::metric_f16::simple_dot::dot_similarity_half;
+use segment::spaces::metric_f16::simple_euclid::euclid_similarity_half;
+use segment::spaces::metric_f16::simple_manhattan::manhattan_similarity_half;
+#[cfg(target_arch = "x86_64")]
+use segment::spaces::metric_f16::sse::cosine::cosine_similarity_sse_half;
+#[cfg(target_arch = "x86_64")]
+use segment::spaces::metric_f16::sse::dot::dot_similarity_sse_half;
+#[cfg(target_arch = "x86_64")]
+use segment::spaces::metric_f16::sse::euclid::euclid_similarity_sse_half;
+#[cfg(target_arch = "x86_64")]
+use segment::spaces::metric_f16::sse::manhattan::manhattan_similarity_sse_half;
 #[cfg(target_arch = "x86_64")]
 use segment::spaces::metric_uint::avx2::cosine::avx_cosine_similarity_bytes;
 #[cfg(target_arch = "x86_64")]
@@ -236,10 +262,169 @@ fn byte_metrics_bench(c: &mut Criterion) {
     });
 }
 
+fn half_metrics_bench(c: &mut Criterion) {
+    let mut group = c.benchmark_group("half-metrics-bench-group");
+
+    let mut rng = StdRng::seed_from_u64(42);
+
+    let random_vectors_1: Vec<Vec<f16>> = (0..COUNT)
+        .map(|_| {
+            (0..DIM)
+                .map(|_| f16::from_f32(rng.gen_range(0.0..=1.0)))
+                .collect()
+        })
+        .collect();
+    let random_vectors_2: Vec<Vec<f16>> = (0..COUNT)
+        .map(|_| {
+            (0..DIM)
+                .map(|_| f16::from_f32(rng.gen_range(0.0..=1.0)))
+                .collect()
+        })
+        .collect();
+
+    group.bench_function("half-dot", |b| {
+        let mut i = 0;
+        b.iter(|| {
+            i = (i + 1) % COUNT;
+            <DotProductMetric as Metric<VectorElementTypeHalf>>::similarity(
+                &random_vectors_1[i],
+                &random_vectors_2[i],
+            )
+        });
+    });
+
+    group.bench_function("half-dot-no-simd", |b| {
+        let mut i = 0;
+        b.iter(|| {
+            i = (i + 1) % COUNT;
+            dot_similarity_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    #[cfg(target_arch = "x86_64")]
+    group.bench_function("half-dot-avx", |b| {
+        let mut i = 0;
+        b.iter(|| unsafe {
+            i = (i + 1) % COUNT;
+            dot_similarity_avx_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    #[cfg(target_arch = "x86_64")]
+    group.bench_function("half-dot-sse", |b| {
+        let mut i = 0;
+        b.iter(|| unsafe {
+            i = (i + 1) % COUNT;
+            dot_similarity_sse_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    #[cfg(target_arch = "aarch64")]
+    group.bench_function("half-dot-neon", |b| {
+        let mut i = 0;
+        b.iter(|| unsafe {
+            i = (i + 1) % COUNT;
+            dot_similarity_neon_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    group.bench_function("half-euclid", |b| {
+        let mut i = 0;
+        b.iter(|| {
+            i = (i + 1) % COUNT;
+            <EuclidMetric as Metric<VectorElementTypeHalf>>::similarity(
+                &random_vectors_1[i],
+                &random_vectors_2[i],
+            )
+        });
+    });
+
+    group.bench_function("half-euclid-no-simd", |b| {
+        let mut i = 0;
+        b.iter(|| {
+            i = (i + 1) % COUNT;
+            euclid_similarity_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    #[cfg(target_arch = "x86_64")]
+    group.bench_function("half-euclid-avx", |b| {
+        let mut i = 0;
+        b.iter(|| unsafe {
+            i = (i + 1) % COUNT;
+            euclid_similarity_avx_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    #[cfg(target_arch = "x86_64")]
+    group.bench_function("half-euclid-sse", |b| {
+        let mut i = 0;
+        b.iter(|| unsafe {
+            i = (i + 1) % COUNT;
+            euclid_similarity_sse_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    #[cfg(target_arch = "aarch64")]
+    group.bench_function("half-euclid-neon", |b| {
+        let mut i = 0;
+        b.iter(|| unsafe {
+            i = (i + 1) % COUNT;
+            euclid_similarity_neon_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    group.bench_function("half-manhattan", |b| {
+        let mut i = 0;
+        b.iter(|| {
+            i = (i + 1) % COUNT;
+            <ManhattanMetric as Metric<VectorElementTypeHalf>>::similarity(
+                &random_vectors_1[i],
+                &random_vectors_2[i],
+            )
+        });
+    });
+
+    group.bench_function("half-manhattan-no-simd", |b| {
+        let mut i = 0;
+        b.iter(|| {
+            i = (i + 1) % COUNT;
+            manhattan_similarity_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    #[cfg(target_arch = "x86_64")]
+    group.bench_function("half-manhattan-avx", |b| {
+        let mut i = 0;
+        b.iter(|| unsafe {
+            i = (i + 1) % COUNT;
+            manhattan_similarity_avx_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    #[cfg(target_arch = "x86_64")]
+    group.bench_function("half-manhattan-sse", |b| {
+        let mut i = 0;
+        b.iter(|| unsafe {
+            i = (i + 1) % COUNT;
+            manhattan_similarity_sse_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+
+    #[cfg(target_arch = "aarch64")]
+    group.bench_function("half-manhattan-neon", |b| {
+        let mut i = 0;
+        b.iter(|| unsafe {
+            i = (i + 1) % COUNT;
+            manhattan_similarity_neon_half(&random_vectors_1[i], &random_vectors_2[i])
+        });
+    });
+}
+
 criterion_group! {
     name = benches;
     config = Criterion::default();
-    targets = byte_metrics_bench
+    targets = byte_metrics_bench, half_metrics_bench
 }
 
 criterion_main!(benches);
