+++ lib/segment/src/spaces/metric_f16/avx/mod.rs
@@ -1,0 +1,3 @@
+pub mod dot;
+pub mod euclid;
+pub mod manhattan;
