+++ lib/segment/src/spaces/metric_f16/mod.rs
@@ -1,10 +1,12 @@
-pub mod simple;
-
-#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
-pub mod simple_sse;
+pub mod simple_dot;
+pub mod simple_euclid;
+pub mod simple_manhattan;
 
 #[cfg(target_arch = "x86_64")]
-pub mod simple_avx;
+pub mod avx;
 
 #[cfg(target_arch = "aarch64")]
-pub mod simple_neon;
+pub mod neon;
+
+#[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+pub mod sse;
